+++ development/http_client/dev_http_client.py
@@ -12,10 +12,7 @@ sys.path.insert(0, APPDIR)
 
 import pyjsonrpc
 
-rpc_client = pyjsonrpc.HttpClient(
-    url = "http://localhost:8080",
-    cookies = {"aa": "AA AA", "bb": "BB"}
-)
+rpc_client = pyjsonrpc.HttpClient("http://localhost:8080")
 
 # Example with *call*
 print repr(rpc_client.call("add", 1, 2))