+++ examples/http_client_usage/simple_example.py
@@ -12,7 +12,7 @@ sys.path.insert(0, APPDIR)
 
 import pyjsonrpc
 
-rpc_client = pyjsonrpc.HttpClient(url = "http://localhost:8080")
+rpc_client = pyjsonrpc.HttpClient("http://localhost:8080")
 
 # Example with *call*
 print rpc_client.call("add", 1, 2)